+++ BetterChests/Framework/Models/DefaultConfig.cs
@@ -1,5 +1,7 @@
 ﻿namespace StardewMods.BetterChests.Framework.Models;
 
+using System.Globalization;
+using System.Text;
 using StardewMods.BetterChests.Framework.Interfaces;
 using StardewMods.BetterChests.Framework.Models.StorageOptions;
 using StardewMods.Common.Services.Integrations.BetterChests.Enums;
@@ -48,4 +50,39 @@ internal sealed class DefaultConfig : IModConfig
 
     /// <inheritdoc />
     public HashSet<string> StashToChestDisableLocations { get; set; } = [];
+
+    /// <inheritdoc />
+    public override string ToString()
+    {
+        StringBuilder sb = new();
+
+        sb.AppendLine(CultureInfo.InvariantCulture, $"{nameof(this.CarryChestLimit)}: {this.CarryChestLimit}");
+        sb.AppendLine(CultureInfo.InvariantCulture, $"{nameof(this.CarryChestSlowLimit)}: {this.CarryChestSlowLimit}");
+        sb.AppendLine(
+            CultureInfo.InvariantCulture,
+            $"{nameof(this.CraftFromChestDisableLocations)}: {string.Join(", ", this.CraftFromChestDisableLocations)}");
+
+        sb.AppendLine(CultureInfo.InvariantCulture, $"{nameof(this.CraftFromWorkbench)}: {this.CraftFromWorkbench}");
+        sb.AppendLine(
+            CultureInfo.InvariantCulture,
+            $"{nameof(this.CraftFromWorkbenchDistance)}: {this.CraftFromWorkbenchDistance}");
+
+        sb.AppendLine(CultureInfo.InvariantCulture, $"{nameof(this.InventoryTabMethod)}: {this.InventoryTabMethod}");
+        sb.AppendLine(CultureInfo.InvariantCulture, $"{nameof(this.LockItem)}: {this.LockItem}");
+        sb.AppendLine(CultureInfo.InvariantCulture, $"{nameof(this.LockItemHold)}: {this.LockItemHold}");
+        sb.AppendLine(CultureInfo.InvariantCulture, $"{nameof(this.SearchItemsMethod)}: {this.SearchItemsMethod}");
+        sb.AppendLine(CultureInfo.InvariantCulture, $"{nameof(this.SearchTagSymbol)}: {this.SearchTagSymbol}");
+        sb.AppendLine(
+            CultureInfo.InvariantCulture,
+            $"{nameof(this.SearchNegationSymbol)}: {this.SearchNegationSymbol}");
+
+        sb.AppendLine(
+            CultureInfo.InvariantCulture,
+            $"{nameof(this.StashToChestDisableLocations)}: {string.Join(", ", this.StashToChestDisableLocations)}");
+
+        sb.AppendLine(CultureInfo.InvariantCulture, $"{nameof(this.DefaultOptions)}: {this.DefaultOptions}");
+        sb.AppendLine(CultureInfo.InvariantCulture, $"{nameof(this.Controls)}: {this.Controls}");
+
+        return sb.ToString();
+    }
 }