+++ BetterChests/Framework/Services/Features/LockItem.cs
@@ -209,10 +209,12 @@ internal sealed class LockItem : BaseFeature<LockItem>
     {
         if (this.IsUnlocked(item))
         {
+            this.Log.Info("{0}: Locking item {1}", this.Id, item.DisplayName);
             item.modData[this.UniqueId] = "Locked";
         }
         else
         {
+            this.Log.Info("{0}: Unlocking item {1}", this.Id, item.DisplayName);
             item.modData.Remove(this.UniqueId);
         }
     }